Abstract out the result executor
This allows specifying custom executors, such as one which executes
on the holder (such as the tile entity or entity)'s tick. This means Sponge
correctly associates lasers with their originating block.

Whilst this is kinda a dirty kludge it works so I'm not going to grumble too much.

See #22