Redo module filtering
This converts IModuleContainer to just being a wrapper for a module set, without
any reference handling.

filterModules now only filters methods on the IModuleContainer implementation,
meaning child methods and meta providers are still applied.

This is slightly more complex to implement, but hopefully makes it easier to use
in a peripheral.