Allow using the mouse inside the keyboard
This fires glasses_* events, which are equivalent to mouse-* events but
operating on the glasses' dimensions.

The system is designed to make it trivial to add support for normal
terminals too (firing mouse_* events for that terminal's dimensions).
However, I wasn't entirely happy with the granularity it provided and so
that feature has been left out for now.

I think this closes #108.