Allow method results to block on arbitrary "resolvers"
Previously methods had one "resolution" strategy: they could wait for
a fixed number of ticks to elapse. We now accept an arbitrary resolver
object, which will be updated every tick, allowing the method to
continue if it returns true.

We offer two default implementations of this resolver - an "immediate"
0-tick delay (for nextThread operations) and a "countdown" resolver (for
operations requiring a delay).