Defer to PlayerInteractionManager as much as possible
This replaces much of the kinetic augment's .use behaviour with that of
PlayerInteractionManager. While the returned values are less useful
(cannot disambiguate between a block's click and item on a block click),
it's less prone to bugs and will hopefully better integrate with core
mods' hacking of the interaction system.

Closes #129.