Add the concept of "object groups" to the overlay glasses
Every object now keeps track of its parent ID (with the glasses having
an ID of 0). The canvas(es) keep a mapping of parent -> children IDs, so
when removing an object we can (recursively) remove all its children.

This implementation may seem a little indirect, but allows us to keep
the objects as plain as possible - which is incredibly useful as we need
to (de)serialise them in order to sync with the client.

Closes #141