Add support for 3d objects
One creates a new "3d canvas", which is centred on the player at that
given moment in time. Any objects added to that canvas exist relative to
the creation point.

We currently only provide a "frame", which allows you to (rather
buggily) draw 2D objects within the 3D world. There's several known
issues with the current implementation:

 - An awful lot of z-fighting
 - Items render "off the page"

One alternative here would be to make this a fixed size, render into a
separate buffer and then blit that to the screen. Something to
investigate in the future perhaps...