Initial implementation of @PlethoraMethod
This attempts to unify most\* IMethod implementations into one concrete
system, as well as simplifying the implementation of method.

@PlethoraMethod effectively acts as a higher-level wrapper over IMethod.
It has a similar usage to to the various @Method.Inject annotations, but
replaces most of them, while meaning you have to write less code. It
offers the following features:

 - You can use IContext<T> or IPartialContext<T>, and we will determine
   the method's type. Otherwise, one must use @FromTarget. This will
   denote the type, and call IPartialContext.getTarget() when invoking
   the method.
 - One can use @FromContext to extract additional data from the context.
   If these are not marked as @Nullable, canApply will ensure their
   presence.
 - While Object[] is still supported, you can specify arguments like any
   other Java call. Primitives are processed using ArgumentHelper,
   everything else is done via an ArgumentType<T> system.
 - One has more flexibility in what you return: primitives are
   automatically boxed, and then all objects are wrapped in an array and
   passed off to MethodResult.result.
 - Module dependencies are part of the annotation.
 - Documentation will hopefully be inferred from the type signature,
   though I haven't looked too much into it yet. It's possible we may
   have to do some hacks in order to get argument names...

This is largely achieved with one super MethodInstance class, and a
dynamic class loader which generates argument validation + context
injection code.

The big advantage here is that many methods no longer need to use
MethodResult.nextTick or the like - as argument validation is done
already, we can do everything in line.