Allow using CC:T networks over XNet
Yes, I said I was going to add this to CC: Tweaked itself.

I lied.

Currently, I'm not 100% sure about this functionalty, and so it seems
safer to keep it in Plethora, even if it could natively be in CC:T. A
bit like RF refueling really.

This registers a channel type, which connects any connected wired nodes
to a central "xnet node".

See #196