Initial stab at a testing server (#206)
This provides a mechanism for starting a Minecraft server with all
supported mods loaded. Once started, we dump documentation and start a
"testing computer". One can then run automated programs on this
computer.

This has a couple of purposes:
 - Generate documentation dumps for all mods, rather than just vanilla
 - Provide a mechanism for automatic testing of Plethora functionality
 - Verify that everything runs on a dedicated server

It would be possible to integrate the testing mechanism into JUnit,
which is what we do for CC:T. However, given that most of the time is
spent actually starting the server, rather than running tests, it
doesn't seem worth it right now.